增加了高频彩最近10期开奖，内蒙快三和值奖金显示，修改了3.10.1bug

+++ Ruyicai_General/main/v_main/src/com/ruyicai/activity/buy/zixuan/AddView.java
@@ -346,8 +346,8 @@ public class AddView {
 		return allZhu;
 	}
 
-	public int getAllAmt() {
-		int allAmt = 0;
+	public long getAllAmt() {
+		long allAmt = 0;
 		for (CodeInfo codeInfo : codeList) {
 			allAmt += codeInfo.getAmt();
 		}
